requirements.txt
Pillow
tk
ttkwidgets
tkhtmlview
dataclasses-json
